Crate detrim

Deserialization trimming for strings in serde models.

Examples

#[derive(Debug, serde::Deserialize)]
struct Form {
    #[serde(deserialize_with = "detrim::string")]
    name: String,
}

let form = serde_json::from_str::<Form>(r#"{ "name": "ferris" }"#).unwrap();
assert_eq!(form.name, "ferris");

let form = serde_json::from_str::<Form>(r#"{ "name": "  ferris   " }"#).unwrap();
assert_eq!(form.name, "ferris");

Functions

cow_str

Trims a CoW string during deserialization.

hashset_stringstd

Trims set of strings during deserialization.

option_string

Trims an optional string during deserialization.

option_string_non_empty

Trims string during deserialization, returning None if it ends up empty.

str

Trims a string slice during deserialization.

string

Trims a string during deserialization.

string_non_empty

Trims string during deserialization, returning error if it ends up empty.

vec_string

Trims list of strings during deserialization.